Total refactor for v13, Slash Commands and confirmation buttons

+++ src/events/reaction-roles.ts
@@ -25,7 +25,9 @@ const ReactionAddEvent: EventTrigger = ['messageReactionAdd', async (messageReac
   if (user.bot) return;
   if (!(await reactionReady(messageReaction))) return;
 
-  const guildId = messageReaction.message.guild!.id;
+  if (!messageReaction.message.guild) return;
+
+  const guildId = messageReaction.message.guild.id;
   const messageId = messageReaction.message.id;
   const member = await messageReaction.message.guild!.members.fetch({
     user,
@@ -101,9 +103,11 @@ const ReactionAddEvent: EventTrigger = ['messageReactionAdd', async (messageReac
 const ReactionRemoveEvent: EventTrigger = ['messageReactionRemove', async (messageReaction: MessageReaction, user: User): Promise<void> => {
   if (user.bot) return;
   if (!(await reactionReady(messageReaction))) return;
-  const guildId = messageReaction.message.guild!.id;
+  if (!messageReaction.message.guild) return;
+
+  const guildId = messageReaction.message.guild.id;
   const messageId = messageReaction.message.id;
-  const member = await messageReaction.message.guild!.members.fetch({
+  const member = await messageReaction.message.guild.members.fetch({
     user,
     force: true,
   });
@@ -125,16 +129,17 @@ const ReactionRemoveEvent: EventTrigger = ['messageReactionRemove', async (messa
 }];
 
 const MessageDeleteEvent: EventTrigger = ['messageDelete', async (message: Message): Promise<void> => {
+  if (!message.guild) return;
   await Promise.all([
     getModels().reaction_roles.destroy({
       where: {
-        guild_id: message.guild?.id,
+        guild_id: message.guild.id,
         message_id: message.id,
       },
     }),
     getModels().reaction_messages_unique.destroy({
       where: {
-        guild_id: message.guild?.id,
+        guild_id: message.guild.id,
         message_id: message.id,
       },
     }),
